moved tests to test_highlevelgraph.py
